feat(ssh,terminal,script,test): close gaps from docs/08
- terminal_view.rs:115: 1行清理,upper.is_ascii_uppercase()
- telnet::resize: 真实 NAWS 子协商 (RFC 1073), connect 时主动 DO NAWS
- session::connect 后台 recv 循环:
  * 接 trigger_engine 注入 (CreateTrigger 现在对 recv 循环立即可见)
  * TerminalBufferUpdated 60Hz 节流 (ticker + dirty 标志)
  * 收 data 后用 from_utf8 调 check_output,命中后 notify_fired + 派发
- script 子系统 0→20 单元测试:
  * trigger_engine: exact/regex/disable/toggle/delete/notify_fired 共 9
  * quick_command: get_text/resolve_targets/current 缺失 错误 共 8
  * sync_input: toggle/subscribe 共 3

附 docs/08-incomplete-features.md 清单(20 项,P0-P3 分级),后续按
该清单继续收尾。

+++ crates/rshell-ui/src/views/terminal_view.rs
@@ -111,9 +111,9 @@ impl TerminalView {
         if has_control {
             if let Some(c) = key_char.and_then(|s| s.chars().next()) {
                 if c.is_ascii_alphabetic() {
-                    let upper = c.to_ascii_uppercase() as u8;
+                    let upper = c.to_ascii_uppercase();
                     if upper.is_ascii_uppercase() {
-                        return vec![upper - b'A' + 1];
+                        return vec![upper as u8 - b'A' + 1];
                     }
                 }
             }
